+++ letsencrypt-win-simple-core/letsencrypt-win-simple-core.csproj
@@ -113,6 +113,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Configuration\Setup.cs" />
+    <Compile Include="Extensions\LifetimeScopeExtensions.cs" />
     <Compile Include="Interfaces\IOptions.cs" />
     <Compile Include="Configuration\Options.cs" />
     <Compile Include="Configuration\Settings.cs" />
